Moved some things out into mappings.go
